# openapi/index.yml
openapi: 3.0.3
info:
  title: FixFlow API
  description: Field Service Management System API
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
...
paths:
...
  /auth/verify-email:
    post:
      tags:
        - Auth

      summary: Verify user email

      description: >
        Verifies the user's email address using the OTP sent during registration.

      operationId: verifyEmail

      requestBody:
        required: true

        content:
          application/json:
            schema:
              type: object

              required:
                - email
                - otp

              properties:
                email:
                  type: string
                  format: email
                  example: omar@example.com

                otp:
                  type: string
                  example: "123456"

      responses:
        "200":
          description: Email verified successfully

          content:
            application/json:
              schema:
                type: object

                properties:
                  success:
                    type: boolean
                    example: true

                  statusCode:
                    type: integer
                    example: 200

                  message:
                    type: string
                    example: Email verified successfully

                  data:
                    type: object
                    example: {}

        "400":
          description: Invalid or expired OTP

        "404":
          description: User not found

        "500":
          description: Internal server error
...
components:
  securitySchemes:

    refreshTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT                    